+++ src/module/applications/hud/token-hud.mjs
@@ -25,6 +25,68 @@ export default class DrawSteelTokenHUD extends foundry.applications.hud.TokenHUD
   /* -------------------------------------------------- */
 
   /** @inheritdoc */
+  _getStatusEffectChoices() {
+    const choiceList = [];
+    for (const status of foundry.utils.iterateValues(CONFIG.statusEffects)) {
+      if ((status.hud === false) || (status.hud?.actorTypes?.includes(this.actor?.type) === false)) {
+        continue;
+      }
+      choiceList.push({
+        _id: status._id,
+        id: status.id,
+        title: _loc(status.name ?? ""),
+        src: status.img,
+        order: status.order ?? 0,
+        isActive: false,
+        isOverlay: false,
+        disabled: false,
+        effects: [],
+      });
+    }
+    const choices = choiceList.sort((a, b) => (a.order - b.order) || a.title.localeCompare(b.title, game.i18n.lang))
+      .reduce((obj, choice) => {
+        obj[choice.id] = choice;
+        return obj;
+      }, {});
+
+    // Update the status of effects which are active for the token actor
+    const activeEffects = this.actor?.effects ?? [];
+    for (const effect of activeEffects) {
+      for (const statusId of effect.statuses) {
+        const status = choices[statusId];
+        if (!status) continue;
+        if (status._id) {
+          if (status._id !== effect.id) {
+            status.disabled = true;
+            status.tooltip = "DRAW_STEEL.ActiveEffect.ComplexStatus";
+          }
+          else status.effects.push(effect.id);
+        } else {
+          if (effect.statuses.size !== 1) {
+            status.disabled = true;
+            status.tooltip = "DRAW_STEEL.ActiveEffect.ComplexStatus";
+          }
+          else status.effects.push(effect.id);
+        }
+        status.isActive = effect.active;
+        if (effect.getFlag("core", "overlay")) status.isOverlay = true;
+      }
+    }
+
+    // Flag status CSS class
+    for (const status of Object.values(choices)) {
+      status.cssClass = [
+        status.isActive ? "active" : null,
+        status.isOverlay ? "overlay" : null,
+      ].filterJoin(" ");
+      status.effectIds = status.effects.join(" ");
+    }
+    return choices;
+  }
+
+  /* -------------------------------------------------- */
+
+  /** @inheritdoc */
   async _onRender(context, options) {
     await super._onRender(context, options);
 
@@ -59,14 +121,19 @@ export default class DrawSteelTokenHUD extends foundry.applications.hud.TokenHUD
       ui.notifications.warn("HUD.WarningEffectNoActor", { localize: true });
       return;
     }
-    const statusId = target.dataset.statusId;
+    const { statusId, effectIds } = target.dataset;
 
-    const effectEnd = this.#effectEnd;
+    const active = !target.classList.contains("active");
+
+    // The condition might exist but be inactive
+    if (active && effectIds) {
+      await this.actor.deleteEmbeddedDocuments("ActiveEffect", effectIds.split(" "));
+    }
 
     await this.actor.toggleStatusEffect(statusId, {
-      active: !target.classList.contains("active"),
+      active,
       overlay: event.button === 2,
-      effectEnd,
+      effectEnd: this.#effectEnd,
     });
   }
 }