+++ src/module/data/models/_module.mjs
@@ -1,3 +1,4 @@
 export { default as ObjectSizeModel } from "./object-size.mjs";
 export { default as SizeModel } from "./size.mjs";
 export { default as SourceModel } from "./source.mjs";
+export * as shapes from "./template-shapes.mjs";